requirements.txt
django
pillow
djangorestframework
django_cors_headers
gunicorn
whitenoise